Move utils.ts from lib/ -> client/
