+++ packages/api/src/client/hostingApi.ts
@@ -1,7 +1,6 @@
 import { Buffer } from 'buffer';
 
 import { createDevLogger } from '../lib/logger';
-import { withRetry } from '../lib/utils';
 import * as domain from '../types';
 import {
   AnalyticsEvent,
@@ -11,6 +10,7 @@ import {
   User,
   getConstants,
 } from '../types';
+import { withRetry } from './utils';
 
 const logger = createDevLogger('hostingApi', false);
